requirements.txt
patch_ng
basicsr @ git+https://git@github.com/StableDraw/BasicSR.git
facexlib @ git+https://git@github.com/StableDraw/facexlib.git
gfpgan @ git+https://git@github.com/StableDraw/GFPGAN.git
numpy
opencv-python
Pillow
torch>=1.7
torchvision
tqdm
wget